chore: cleaned up hooks
